Limit get_string() length to string descriptor length
Use the length as reported by the first byte of the descriptor.
Otherwise (at least for some backend/kernel combinations) the buffer
length provided by the user would be used instead, and NUL characters
would be appended to the returned string.

This also makes the legacy getString() function compatible with
pyusb 0.4.