Fixed compatibility with Python 3.3. closes #8.
In usb/_interop.py in the as_array function, the array typecodes 'c' and
'u' are used. The 'c' typecode was removed from python 3.3 and 'u' will
is deprecated. The solution is to use the fromstring array method, but
it has the side effect of adding only the first byte of Unicode strings.