Fix: use the context parameter across libusb functions. closes #27
To avoid double initialization when PyUSB is used with another library
thats also uses libusb, we started to make use of the context libusb
argument, instead of using the default one, but we didn't use it to call
several libusb functions, which caused segfaults on some versions of
libusb. This commit fixes this.