Documentation update.
From USB spec point of view, you don't need to know the interface number
of an endpoint to issue a transfer on that endpoint, given you know the
endpoint address. This is true because endpoint addresses must be unique
across interfaces. Nevertheless, some backends may need to know the
interface number.