Fix braino in kernel driver attachement methods.
get_interface was removed in the commit 20cd8c because it was not
necessary to I/O methods to know the interface number, but it was still
referenced in the kernel driver attachment methods.

This commit gets rid of get_interface by requiring the library use pass
the interface number to these methods.

closes #37.