Allow inplace read transfers.
Now the read function can receive an output buffer of the type array
where the data transfered will be put in. This allows the buffer to be
reused for future read transfers, avoiding the overhead of creating a
new object for each call.

This is based on the original work of Marijn van Vliet (pull
request #25), but instead of creating a new function readinto, we
"overload" the read function to allow it to receive the buffer.