libusb1: fix hanging when the device is disconnected in a iso transfer.
If the device is disconnected when a isochronous transfer is ongoing,
it might raise an exception in the callback and the 'callback done'
signaling flag is never set. This causes the program to hang
forever.

We move the error handling to the caller function, after the callback
returns. In this way, the exception is raised in the main thread and
client software is notified about the exceptional situation.

closes #85.