core: Fix seg fault when dispose is called after device disconnection.
When dispose is called after the device is diconnected,
release_interface might fail (see commit 5e9ef1536), then the interface
stays in the claimed interfaces set. When the device is destroyed, its
__del__ method will call dispose again and, since the interface is
still marked as claimed, a new call to release_interface will be made
with an invalid device handle, causing a segmentation fault.

We fix that by always removing the interface from the claimed set.

closes #89.