Avoid hardcoding the index for _PC_PATH_MAX.
POSIX explicitly leaves the values of these named constants unspecified.
The value 4 is right for linux but wrong for solaris. Who knows what it is
in another OS. The result of having the wrong index is the wrong value
gets queried, making the ctypes structures the wrong sizes and causing
pyusb to fail in perplexing ways.

It's better to use the name ('PC_PATH_MAX') and allow Python to use the
correct index for the platform.