Add a hardcoded PATH_MAX for Solaris.
The preceding commit would make it work in most cases, but perhaps it
is best not to rely on pathconf when the value from limits.h is known.
In theory, pathconf could return a different value depending on the
filesystem where . is, but the fixed value in limits.h will be what
libusb was compiled with.