Expose device's LANGIDs; suppress missing strings.
When a device does not have a certain string (e.g. serial number),
it supplies zero as the index for that string. Zero is never the
index of a real string, only a signal not to request it. A request
for string index zero has a special meaning, to find the list of
languages the device strings are offered in. When code requested a
string without testing the index for zero, it could get back a bogus
string like a Cyrillic Љ character, as a result of trying to interpret
the language ID codes as Unicode characters.

Now get_string simply returns None if passed an index of zero.

At the same time, there was no simple way for calling code to examine
the supported language IDs for the device. The spec forbids sending
string requests for a language the device does not support, so the
supported IDs should be easy for the caller to check. Now there is
a langids property on Device.

If a device has no strings, the spec also allows it to have no langids
list. The device might reply with an empty list, or fail the request
with a backend pipe error. The latter is caught in the property accessor
on core.Device, rather than in util.get_langids, because of an order
constraint between when core imports util and when it defined USBError.

The string property accessors on Device normally call get_string once
and cache the result, but if the index is zero (so the result is None)
they will end up calling get_string every time ... but all it does in
that case is if 0 == index: return None so the cost is negligible and
the code stays simple.