Raise a USBError for obviously invalid descriptors
In certain scenarios some devices may return zero-length descriptors.
However, all descriptors must have at least two bytes (the descriptor
length in bytes and the descriptor type), and the USB 3.2 specification
states that:

> If a descriptor returns with a value in its length field that is less
> than defined by this specification, the descriptor is invalid and should
> be rejected by the host.

Instead of returning invalid descriptors, modify get_descriptor so it
raises a proper USBError as soon as/if the descriptor is found to be
invalid.

This results in get_string (which depends on get_descriptor) no longer
failing with a generic IndexError; instead, the user will be able to
catch/handle the USBError from the underlying call.

Fixes: #164 ('"IndexError: array index out of range" when reading
product property')
Reported-by: @c1728p9