Normalize all copyright disclaimers
PyUSB 1.0 was originally licensed in 2010 under the Tcl/Tk License
Terms.[1][2]

Shortly after the original author changed the license to the 3-clause
BSD, but at the same time added copyright notices to each source file
that were still based on the previous Tcl/Tk terms.[3][4]

The license was once again changed in 2018 by a (at the time) new
maintainer.  The Apache 2.0 license was chosen, and all copyright
notices where adjusted accordingly.[5]

However, this was a more restrictive license than before that, in
particular, created compatibility issues with GPL 2.0-only projects.
The change was ultimately reverted after a brief discussion, but
copyright notices and metadata where again left out-of-sync with the
license.[6][7]

Fix all copyright disclaimers and clarify that all code is currently
licensed under the BSD 3-clause license, as the original author had
intended back in 2010.

[1] 270bd71559e9 ("[1.0] Added README, LICENSE, TODO and ACKNOWLEDGEMENTS files.")
[2] https://www.tcl.tk/software/tcltk/license.html
[3] 0f30969ec3ac ("[1.0] License update.")
[4] fa827adb61e1 ("Updated license information for test scripts.")
[5] 161086acdf9c ("Moving to the standard Apache 2.0 License.")
[6] 9d6d09460370 ("back to bsd 3 clause")
[7] #244 ("GitHub thinks this is Apache 2.0?")