tree: guard finalizers against aborted initialization
AutoFinalizedObject sets up the finalizer to run using
weakref.finalize() within the object allocation – __new__().  This means
that the finalizer is set up to run regardless of whether object
initialization – __init__() – completed successfully.

If __init__() raises an exception, or if it is aborted due to an
external reason (e.g. process termination due to an exception on another
thread or a KeyboardInterrupt), the finalizer will run with a partially
constructed object.

Therefore, check the presence (and, when applicable, the validity) of
any fields of self used by the finalizer before they are actually used.

Related: https://github.com/pyusb/pyusb/issues/380#issuecomment-1084140904