util: raise TypeError if create_buffer doesn't receive a length
`Device.ctrl_transfer` depends on `create_buffer` raising a `TypeError`
if its parameter is not a valid length. In particular, if an array is
passed in, `create_buffer` must *not* return a copy of it.

Commit 3ea79b01af08 accidentally broke this and, consequently, any calls
to `ctrl_transfer` passing in a buffer. Thanks @kauwua for reporting the
issue and discussing the potential solutions.

Note that `length * b'\x00'` is actually faster than `bytes(length)`
(tested with Python 3.13 on x86_64).

Fixes: #528
Fixes: 3ea79b01af08 ("util: remove left over python 2 compatibility workaround")