interop: prioritize no-op case but also accept length for a new array
While discussing #527, @kauwua made the valid point that we could
optimize for the case where the user has passed in a byte `array`.

In that spirit, prioritize the no-op conversion case, but also accept
the parameter to be a length for a new array in preparation for
replacing the conditional in `Device.ctrl_transfer`.

Related: #528